backend/requirements.txt
langchain
langchain-community
python-dotenv
faiss-cpu
pypdf
sentence-transformers
langchain-groq
tavily-python

fastapi
uvicorn
python-multipart